guardar detalles

+++ src/main/java/com/portfolio/CynthiaMC/Repository/RSkills.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface RSkills extends JpaRepository<Skills, Long> {
+public interface RSkills extends JpaRepository<Skills, Integer> {
      public Optional<Skills> findByNombreE(String nombreE);
     public boolean existsByNombreE(String nombreE);
 }
