+++ src/main/java/com/portfolio/CynthiaMC/Security/Repository/iRolRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface iRolRepository extends JpaRepository<Rol, Long> {
+public interface iRolRepository extends JpaRepository<Rol, Integer> {
     Optional<Rol> findByRolNombre(RolNombre rolNombre);
     
 }
